adk-agents/tauri_frontend_agent/requirements.txt
# Tauri Frontend Agent - Dependencies
# =====================================

# ADK Core
google-adk>=1.0.0
google-genai>=0.1.0

# FastAPI server
fastapi>=0.100.0
uvicorn>=0.20.0

# Pydantic para validacao
pydantic>=2.0.0

# Environment variables
python-dotenv>=1.0.0
